backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
PyJWT==2.8.0
python-dotenv==1.0.0
numpy==1.24.3
openai==1.3.7
pytest==7.4.3
pytest-asyncio==0.21.1
